backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.11.0
asttokens==3.0.1
attrs==25.4.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
decorator==5.2.1
defusedxml==0.7.1
distro==1.9.0
Django==6.0.1
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
djoser==2.3.3
executing==2.2.1
frozenlist==1.8.0
google-auth==2.47.0
google-genai==1.59.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
ipython==8.38.0
jedi==0.19.2
matplotlib-inline==0.2.1
multidict==6.7.0
oauthlib==3.3.1
openai==0.27.10
packaging==25.0
parso==0.8.5
pexpect==4.9.0
prompt_toolkit==3.0.52
propcache==0.4.1
psycopg==3.3.2
psycopg-binary==3.3.2
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.10.1
python-dotenv==1.0.1
python3-openid==3.2.0
regex==2026.1.15
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
sniffio==1.3.1
social-auth-app-django==5.7.0
social-auth-core==4.8.3
sqlparse==0.5.5
stack-data==0.6.3
tabulate==0.9.0
tenacity==9.1.2
tiktoken==0.3.3
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
wcwidth==0.2.14
websockets==15.0.1
yarl==1.22.0
